+++ client/jsconfig.json
@@ -1,0 +1,6 @@
+{"compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["./src/*"]
+    }
+  }}
