+++ client/src/Features/NewRecipe/index.js
@@ -1,0 +1,1 @@
+export {CreateNewRecipe} from './components/CreateNewRecipe'
