+++ client/src/Features/NewRecipe/utils/createBlobFromDataURL.js
@@ -1,0 +1,16 @@
+// Create Blob file from URL
+  // ref: https://www.geeksforgeeks.org/how-to-convert-data-uri-to-file-then-append-to-formdata/
+
+  export function createBlobFromDataURL(inputURI) {
+    var byteString = window.atob(inputURI.split(',')[1]);
+    var mimeString = inputURI.split(',')[0].split(':')[1].split(';')[0]
+    var ab = new ArrayBuffer(byteString.length);
+    var ia = new Uint8Array(ab);
+  
+    for (var i = 0; i < byteString.length; i++) {
+      ia[i] = byteString.charCodeAt(i);
+    }
+  
+    var blob = new Blob([ab], {type: mimeString});
+    return blob;
+  }
