--- client/src/jsconfig.json
@@ -1,8 +1,0 @@
-{
-    "compilerOptions": {
-        "baseUrl": ".",
-        "paths": {
-        "@/*": ["./src/*"]
-        }
-    }
-}
